+++ x-pack/heartbeat/monitors/browser/synthexec/synthexec.go
@@ -15,7 +15,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
 	"runtime"
 	"strings"
 	"sync"
@@ -345,29 +344,6 @@ func lineToSynthEventFactory(typ string) func(bytes []byte, text string) (res *S
 			},
 		}, nil
 	}
-}
-
-var emptyStringRegexp = regexp.MustCompile(`^\s*$`)
-
-// jsonToSynthEvent can take a line from the scanner and transform it into a *SynthEvent. Will return
-// nil res on empty lines.
-func jsonToSynthEvent(bytes []byte, text string) (res *SynthEvent, err error) {
-	// Skip empty lines
-	if emptyStringRegexp.Match(bytes) {
-		return nil, nil
-	}
-
-	res = &SynthEvent{}
-	err = json.Unmarshal(bytes, res)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if res.Type == "" {
-		return nil, fmt.Errorf("unmarshal succeeded, but no type found for: %s", text)
-	}
-	return res, err
 }
 
 // getNpmRoot gets the closest ancestor path that contains package.json.
